requirements.txt
Pillow
ttkbootstrap
requests
